backend/services/api/requirements.txt
# API Framework
fastapi>=0.95.0
uvicorn>=0.21.0
starlette>=0.27.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.27.0
alembic>=1.9.0

# Message broker
confluent-kafka>=2.0.0

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
websockets>=11.0.0
email-validator>=2.0.0

# Monitoring
prometheus-client>=0.16.0
opentelemetry-api>=1.16.0
opentelemetry-sdk>=1.16.0
opentelemetry-instrumentation-fastapi>=0.38.0

# Caching
redis>=4.5.0 